requirements.txt
inquirer==3.1.2
openai==1.11.1
python-dotenv==0.21.1
google-generativeai==0.3.1
